requirements.txt
alembic==1.2.0
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.3.0
autopep8==1.4.4
black==19.3b0
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
docopt==0.6.2
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
evdev==0.7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Guake==3.0.5
gunicorn==19.9.0
httpie==0.9.8
httplib2==0.9.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
jsonpickle==1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
louis==3.5.0
lutris==0.5.4
macaroonbakery==1.1.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
netifaces==0.10.4
notify2==0.3
numpy==1.17.2
oauth==1.0.1
olefile==0.45.1
pandas==0.25.1
passlib==1.7.1
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
pipreqs==0.4.9
protobuf==3.0.0
psutil==5.6.3
psycopg2-binary==2.7.5
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.4.2
pygobject==3.26.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pyte==0.8.0
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
system-service==0.3
systemd-python==234
thefuck==3.29
toml==0.10.0
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.6
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
xkit==0.0.0
yarg==0.1.9
zope.interface==4.3.2
